moved db methods to helper.py
