+++ quickannotator/db/helper.py
@@ -1,0 +1,16 @@
+from quickannotator.db import Base, db
+from sqlalchemy import Table
+
+
+def create_dynamic_model(table_name, base=Base):
+    class DynamicAnnotation(base):
+        __tablename__ = table_name
+        __table__ = Table(table_name, base.metadata, autoload_with=db.engine)
+
+    return DynamicAnnotation
+
+
+def build_annotation_table_name(image_id: int, annotation_class_id: int, is_gt: bool):
+    gtpred = 'gt' if is_gt else 'pred'
+    table_name = f"annotation_{image_id}_{annotation_class_id}_{gtpred}"
+    return table_name
